Major Polarimetry reform

+++ SigLib.py
@@ -43,7 +43,7 @@ import Util
 class SigLib:
     def __init__(self):       
         self.cfg = os.path.expanduser((sys.argv[1]))
-
+        print(self.cfg)
         #config = ConfigParser.RawConfigParser()
         config = RawConfigParser() # Needs to be tested for python2 compatibility 
         config.read(self.cfg)
@@ -86,6 +86,11 @@ class SigLib:
 
         self.issueString = ""
         self.zipname = None
+        self.unzipdir = None
+        self.fname = None
+        self.sattype = None
+        self.granule = None
+        self.imgname = None
         self.count_img = 0            # Number of images processed
         self.bad_img = 0             # Number of bad images processed
         self.starttime = strftime('%Y%m%d_%H%M%S', localtime())
@@ -93,7 +98,8 @@ class SigLib:
             self.starttime +'.cfg')) # make a copy of the cfg file
         self.length_time = 0
         self.loghandler = None
-        self.logger = 0        
+        self.logger = 0
+        self.sar_meta = None
     
     def createLog(self,zipfile=None):   
         """
@@ -118,10 +124,9 @@ class SigLib:
         self.logger.propagate = False
                 
         self.logger.info("SigLib Run w/ config: %s", self.cfg)
-        self.logger.info("User: %s",os.getenv('USER'))        
+        self.logger.info("User: %s",os.getenv('USER'))
 
-
-    def proc_File(self, zipfile):
+    def proc_File(self, zipfile, cleanup=True):
         """
         Locates a single satellite image zip file and processes it according 
         to the config file.  Note this cannot be nested in proc_dir since the 
@@ -132,41 +137,30 @@ class SigLib:
             *zipfile* 
         """      
 
-        self.logger = self.createLog(zipfile)
-        self.logger = logging.getLogger(__name__)
         self.count_img = 1      
                            
         self.logger.info('Started processing image %s', zipfile) 
         
         try:
-            start_time = time.time()
             self.retrieve(zipfile)
             self.logger.debug('image retrieved')
-            # Do clean-up
                             
         except Exception as e: #Normally Exception, e
             self.logger.error('Image failed %s, due to: %s', zipfile, e, exc_info=True)
             self.logger.error("Image processing exception, moving to next image")
-            self.issueString += "\n\nERROR (exception): " + zipfile
             self.bad_img = 1
-    
-        end_time = time.time()
-        self.logger.info("Image Processing Time: " + str(int((end_time-start_time)/60)) + " Minutes " + str(int((end_time-start_time)%60)) + " Seconds")
-        os.chdir(self.tmpDir)
 
-        try:
-            shutil.rmtree(os.path.splitext(os.path.basename(zipfile))[0])
-        except Exception as e:
-            self.logger.debug("Warning: could not remove file from temp directory; {}".format(e))
-        self.logger.debug('cleaned zip dir')
+        if cleanup:
+            os.chdir(self.tmpDir)
+            try:
+                shutil.rmtree(os.path.splitext(os.path.basename(zipfile))[0])
+            except Exception as e:
+                self.logger.debug("Warning: could not remove file from temp directory; {}".format(e))
+            self.logger.debug('cleaned zip dir')
 
         good_img = self.count_img - self.bad_img
         self.logger.info("%i images were successfully processed out of %i", good_img, self.count_img)
 
-        if self.bad_img > 0:
-            # Write the issue file
-            self.logger.error(self.issueString)
-          
     def proc_Dir(self, path, pattern):
         """
         Locates satelite image raw data files (zipfiles) using a
@@ -205,18 +199,14 @@ class SigLib:
             self.count_img += 1
             
             try:
-                start_time = time.time()
                 self.retrieve(zipfile)
                 self.logger.debug('image retrieved') #TODO move to retrieve and return meaningful info.
             except Exception as e: #Normally Exception
                 self.logger.error('Image failed %s, due to: %s', zipfile, e, exc_info=True)
                 self.logger.error("Image processing exception, moving to next image")
-                self.issueString += "\n\nERROR (exception): " + zipfile
                 self.bad_img += 1
                   
             # Do clean-up
-            end_time = time.time()
-            self.logger.info("Image Processing Time: " + str(int((end_time-start_time)/60)) + " Minutes " + str(int((end_time-start_time)%60)) + " Seconds")
             os.chdir(self.tmpDir)
             try:
                 shutil.rmtree(os.path.splitext(os.path.basename(self.zipname))[0])
@@ -226,17 +216,13 @@ class SigLib:
 
         good_img = self.count_img - self.bad_img
         self.logger.info("%i images were successfully processed out of %i", good_img, self.count_img)
-
-        if self.bad_img > 0:
-            # Write the issue file
-            self.logger.error(self.issueString)
             
-        self.logger.handlers = []
-        logging.shutdown()
-        del sys.modules['Image']
-        del sys.modules['Metadata']
-        del sys.modules['Database']
-        del sys.modules['Util']
+        #self.logger.handlers = []
+        #logging.shutdown()
+        #del sys.modules['Image']
+        #del sys.modules['Metadata']
+        #del sys.modules['Database']
+        #del sys.modules['Util']
 
     def retrieve(self, zipfile):
         """
@@ -251,26 +237,33 @@ class SigLib:
         # Verify if zipfile has its own subdirectory before unzipping
         unzipdir, zipname, nested, granule = Util.getZipRoot(os.path.join(self.scanDir,zipfile), self.tmpDir)
 
-        self.zipname = zipname #for file cleanup
+        self.zipname = zipname
+        self.unzipdir = unzipdir
+        self.granule = granule
+
         self.logger.debug("Zipfile %s will unzip to %s. Granule is %s and Nested is %s", zipfile, unzipdir, granule, nested)        
 
         # Unzip the zip file into the unzip directory
         Util.unZip(zipfile, unzipdir)
         self.logger.debug("Unzip ok")
         
-        if unzipdir == self.tmpDir:      # If files have been unzipped in their own subdirectory
-            unzipdir = os.path.join(self.tmpDir, zipname)    # Then correct the name of unzipdir
+        if self.unzipdir == self.tmpDir:      # If files have been unzipped in their own subdirectory
+            self.unzipdir = os.path.join(self.tmpDir, self.zipname)    # Then correct the name of unzipdir
             if nested == 1:   # If zipfile has nested directories
-                unzipdir = os.path.join(unzipdir, zipname)    # Then correct the name of unzipdir
+                self.unzipdir = os.path.join(self.unzipdir, self.zipname)    # Then correct the name of unzipdir
                 
         # Parse zipfile
-        fname, imgname, sattype = Util.getFilename(granule, unzipdir, self.loghandler)
+        fname, imgname, sattype = Util.getFilename(self.granule, self.unzipdir, self.loghandler)
+
+        self.fname = fname
+        self.imgname = imgname
+        self.sattype = sattype
 
         if sattype == 'SEN-1':
-            granule = granule.split('.')[0]
-            zipname = granule
-            os.rename(unzipdir, os.path.join(self.tmpDir, granule))
-            unzipdir = os.path.join(self.tmpDir, granule)
+            self.granule = self.granule.split('.')[0]
+            self.zipname = self.granule
+            os.rename(self.unzipdir, os.path.join(self.tmpDir, self.granule))
+            self.unzipdir = os.path.join(self.tmpDir, self.granule)
         
         formatter = logging.Formatter('')        
         self.loghandler.setFormatter(formatter)
@@ -281,15 +274,15 @@ class SigLib:
         formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')         
         self.loghandler.setFormatter(formatter)
         
-        if fname == "error":
+        if self.fname == "error":
             self.logger.error("File not valid or available, moving to next file")
             self.issueString += "\n\nERROR (retrieve): " + zipfile    # Take note
             self.bad_img += 1
 
         else:#begin processing data ...
-            sar_meta = func_timeout(300, Metadata, args=(granule, imgname, unzipdir, zipfile, sattype, self.loghandler))   # Retrieve metadata          
+            self.sar_meta = func_timeout(300, Metadata, args=(self.granule, self.imgname, self.unzipdir, zipfile, self.sattype, self.loghandler))   # Retrieve metadata
 
-            if sar_meta.status != "ok":       # Meta class unsuccessful
+            if self.sar_meta.status != "ok":       # Meta class unsuccessful
                 self.logger.error("Creating an instance of the meta class failed, moving to next image")
                 self.bad_img += 1
                 self.issueString += "\n\nERROR (meta class): " + zipfile
@@ -297,31 +290,37 @@ class SigLib:
             else:
                 formatter = logging.Formatter('')        
                 self.loghandler.setFormatter(formatter)
-                self.logger.info('\n' + 'Dimgname: %s', sar_meta.dimgname + '\n')
+                self.logger.info('\n' + 'Dimgname: %s', self.sar_meta.dimgname + '\n')
                 formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')         
                 self.loghandler.setFormatter(formatter)
 
-            sar_meta.saveMetaFile(self.imgDir)
+            #sar_meta.saveMetaFile(self.imgDir)
 
-            #"Utility" function, to be replaced by Query            
-            if self.processData2db == "1":
-                db = Database(self.table_to_query, self.dbName, loghandler=self.loghandler, host=self.dbHost)  # Connect to the database
-                self.data2db(sar_meta, db, zipfile)
-                db.removeHandler()
+            #"Utility" function, to be replaced by Query
+            if "1" in {self.processData2db, self.qualitativeProcess, self.quantitativeProcess}:
+                start_time = time.time()
 
-            if self.qualitativeProcess == "1":
-                self.logger.debug("processing data to image")
-                self.qualitative_mode(fname, imgname, zipname, sattype, granule, zipfile, sar_meta, unzipdir)
-                
-            if self.quantitativeProcess == "1":
-                db = Database(self.table_to_query, self.dbName, loghandler=self.loghandler, host=self.dbHost)
-                self.quantitative_mode(db, fname, imgname, zipname, sattype, granule, zipfile, sar_meta, unzipdir)
-                db.removeHandler()
-                
-        return zipname #for temp folder cleanup
+                if self.processData2db == "1":
+                    db = Database(self.table_to_query, self.dbName, loghandler=self.loghandler, host=self.dbHost)  # Connect to the database
+                    self.data2db(db, zipfile)
+                    db.removeHandler()
+
+                if self.qualitativeProcess == "1":
+                    self.logger.debug("processing data to image")
+                    self.qualitative_mode(fname, imgname, zipname, sattype, granule, zipfile, unzipdir)
+
+                if self.quantitativeProcess == "1":
+                    db = Database(self.table_to_query, self.dbName, loghandler=self.loghandler, host=self.dbHost)
+                    self.quantitative_mode(db, fname, imgname, zipname, sattype, granule, zipfile, unzipdir)
+                    db.removeHandler()
+
+                end_time = time.time()
+
+                self.logger.info("Image Processing Time: " + str(int((end_time - start_time) / 60)) + " Minutes " + str(
+                    int((end_time - start_time) % 60)) + " Seconds")
 
     #This will be moved to a utility? rather than a mode        
-    def data2db(self, meta, db, zipfile):
+    def data2db(self, db, zipfile):
         """
         Adds the image file metadata to tblmetadata table in the specified database.
         Will create/overwrite the table tblmetadata if prompted (be carefull)
@@ -335,8 +334,8 @@ class SigLib:
         #TODO: implement ROI filtering such that only images within the ROI
         #are uploaded to tblmetadata
         print("Starting data2db")
-        if meta.status == "ok":
-            meta_dict = meta.createMetaDict()  # Create dictionary of all the metadata fields
+        if self.sar_meta.status == "ok":
+            meta_dict = self.sar_meta.createMetaDict()  # Create dictionary of all the metadata fields
             db.meta2db(meta_dict)       # Upload metadata to database
           
         else:
@@ -363,7 +362,7 @@ class SigLib:
         pass
             
 
-    def qualitative_mode(self, fname, imgname, zipname, sattype, granule, zipfile, sar_meta, unzipdir):
+    def qualitative_mode(self, fname, imgname, zipname, sattype, granule, zipfile, unzipdir):
         """
         Opens an image file and converts it to the format given in the config file
 
@@ -396,7 +395,7 @@ class SigLib:
         os.chdir(newTmp)
             
         # Process the image
-        sar_img = func_timeout(600, Image, args=(fname, unzipdir, sar_meta, self.imgType, self.imgFormat, zipname, self.imgDir, newTmp, self.loghandler, self.elevation_correction))
+        sar_img = func_timeout(600, Image, args=(fname, unzipdir, self.sar_meta, self.imgType, self.imgFormat, zipname, self.imgDir, newTmp, self.loghandler, self.elevation_correction))
 
         
         if sar_img.status == "error":
@@ -450,11 +449,11 @@ class SigLib:
             #sar_img.cleanFiles(levels=['nil','proj'])
             #self.logger.debug('Intermediate file cleanup done')
             sar_img.removeHandler()
-            sar_meta.removeHandler()
+            self.sar_meta.removeHandler()
         print("Quatlitative Mode Complete.")
     
                  
-    def quantitative_mode(self, db, fname, imgname, zipname, sattype, granule, zipfile, sar_meta, unzipdir):          
+    def quantitative_mode(self, db, fname, imgname, zipname, sattype, granule, zipfile, unzipdir):
         """
         Process images quantitative_modely, based on an ROI in the database, and per zipfile:
             -Qry to find what polygons in the ROI overlap this image
@@ -483,7 +482,7 @@ class SigLib:
         newTmp = os.path.join(self.tmpDir,zipname)
         
         # Process the image
-        sar_img = func_timeout(600, Image, args=(fname, unzipdir, sar_meta, self.imgType, self.imgFormat, zipname, self.imgDir, newTmp, self.loghandler))
+        sar_img = func_timeout(600, Image, args=(fname, unzipdir, self.sar_meta, self.imgType, self.imgFormat, zipname, self.imgDir, newTmp, self.loghandler))
 
         if sar_img.status == "error":
             self.logger.error("Image could not be opened or manipulated, moving to next image")
@@ -548,8 +547,7 @@ class SigLib:
                     db.imgData2db(imgData, bandName, inst, sar_img.meta.dimgname, zipname)
             else:
                 stats = sar_img.getImgStats(save_stats = True)
-                #db.stats2db(stats, inst, granule, self.roi)
-                sar_img.applyStretch(stats, procedure='std', sd=3, sep=sep, inst=inst)   
+                sar_img.applyStretch(stats, procedure='std', sd=3, sep=sep, inst=inst)
             sar_img.cleanFiles(levels=['proj', 'crop'])
             
         sar_img.cleanFiles(levels=['nil']) 
@@ -582,6 +580,8 @@ class SigLib:
         if self.scanPath == "1":
             self.proc_Dir(self.scanDir, self.scanFor)      # Scan by path pattern
         elif self.scanFile == "1":
+            self.logger = self.createLog(os.path.abspath(os.path.expanduser(str(sys.argv[-1]))))
+            self.logger = logging.getLogger(__name__)
             self.proc_File(os.path.abspath(os.path.expanduser(str(sys.argv[-1]))))  #assume this is the last arg (after 'jobid')
         else:
             print("\nPlease specify one method to scan the data in the config file.\n")
